@@ -19,7 +19,7 @@
     "sort-packages": true
   },
   "require": {
-    "php": "^7.2 || ^8.0",
+    "php": "^7.4 || ^8.0",
     "beberlei/assert": "^3.2",
     "nikolaposa/rate-limit": "^2.0",
     "psr/http-server-middleware": "^1.0"
